Make TimeEvolution module an extention
