+++ ext/TimeEvolution/FFT_analysis.jl
@@ -1,3 +1,4 @@
+import HarmonicBalance: FFT, FFT_analyze, u_of_t, uv_nonrotating_frame
 export FFT, u_of_t, uv_nonrotating_frame
 
 """Fourier transform the timeseries of a simulation in the rotating frame and calculate the quadratures and freqeuncies in the non-rotating frame."""
