+++ ext/TimeEvolution/TimeEvolution.jl
@@ -1,0 +1,18 @@
+module TimeEvolution
+
+    using HarmonicBalance
+    using Symbolics
+    using Plots
+    using OrdinaryDiffEq
+    using OrderedCollections
+    using DSP
+    using FFTW
+    using Peaks
+    using DocStringExtensions
+
+    include("sweeps.jl")
+    include("ODEProblem.jl")
+    include("FFT_analysis.jl")
+    include("hysteresis_sweep.jl")
+
+end
